+++ .eslintrc.json
@@ -13,6 +13,7 @@
   "plugins": ["@typescript-eslint"],
   "rules": {
     "no-console": "off",
-    "import/prefer-default-export": "off"
+    "import/prefer-default-export": "off",
+    "no-plusplus": "off"
   }
 }
