+++ package.json
@@ -7,6 +7,7 @@
   "scripts": {
     "start:prod": "ts-node app.ts",
     "start:dev": "nodemon app.ts",
+    "start:multi": "ts-node multi.ts",
     "check:eslint": "eslint . --ext .ts --fix && prettier --write ."
   },
   "keywords": [],
@@ -14,6 +15,7 @@
   "license": "ISC",
   "dependencies": {
     "@types/uuid": "^8.3.4",
+    "dotenv": "^16.0.1",
     "nodemon": "^2.0.16",
     "ts-node": "^10.8.1",
     "uuid": "^8.3.2"
